Issue #839 Refactor to move service search data validation to its own top level folder
