+++ src/async/location.ts
@@ -1,7 +1,7 @@
 import * as Permissions from 'expo-permissions';
 import * as Location from 'expo-location';
 import { Errors } from '../errors/types';
-import { LatLong } from '../components/search/types';
+import { LatLong } from '../validation/search/types';
 
 export type NoLocationPermissionErrorAction = Readonly<ReturnType<typeof noLocationPermissionError>>;
