+++ src/components/help/help_connected_component.tsx
@@ -11,7 +11,7 @@ import {
     setManualUserLocation,
 } from '../../stores/manual_user_location';
 import { selectManualUserLocation } from '../../selectors/services/select_manual_user_location';
-import { LatLong } from '../search/types';
+import { LatLong } from '../../validation/search/types';
 
 const mapStateToProps = (store: Store, ownProps: RouterProps): HelpComponentProps => ({
     history: ownProps.history,
