+++ src/components/search/api/configuration.ts
@@ -1,5 +1,5 @@
 // tslint:disable:no-expression-statement
-import { LatLong } from '../types';
+import { LatLong } from '../../../validation/search/types';
 
 export const toServiceSearchConfiguration = (latlong?: LatLong): Object => {
     const latLongArgument = latlong ? { aroundLatLng: toAlgoliaParameter(latlong) } : {};
