+++ src/selectors/services/select_manual_user_location.ts
@@ -1,5 +1,5 @@
 import { Store } from '../../stores';
-import { LatLong } from '../../components/search/types';
+import { LatLong } from '../../validation/search/types';
 
 export const selectManualUserLocation = (store: Store): LatLong | undefined => (
     store.manualUserLocation.userLocation
