+++ src/stores/services/actions.ts
@@ -3,7 +3,7 @@ import * as constants from '../../application/constants';
 import * as helpers from '../helpers/make_action';
 import { Service } from './types';
 import { Errors } from '../../errors/types';
-import { LatLong } from '../../components/search/types';
+import { LatLong } from '../../validation/search/types';
 
 export type BuildTopicServicesRequestAction = Readonly<ReturnType<typeof buildTopicServicesRequestAction>>;
