+++ src/validation/search/types.ts
